% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.spfda.R
\name{diagnostic}
\alias{diagnostic}
\title{Diagnostic the result of Spatial scalar on functional regression}
\usage{
diagnostic(object, which = "beta", bidx = 1)
}
\arguments{
\item{object}{object of class \code{spfda}.}

\item{which}{Select parameter for mcmc diagnostic. "beta", "sigma2", "lambda" are available.}

\item{bidx}{vector of integer, Index of beta for checking MCMC chain convergence.}
}
\value{
\code{diagnostic} returns plots for checking MCMC convergence.
}
\description{
\code{diagnostic} is used to diagnostic the result of Spatial scalar on functional regression.
}
