% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fspreg-package.R
\name{fspreg}
\alias{fspreg}
\title{fspreg: A Package for Spatial scalar on functional regression}
\description{
The `fspreg` package provides tools and functions for Spatial scalar on functional regression.
}
