% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.spfda.R
\name{plot.spfda}
\alias{plot.spfda}
\title{Plotting the result of Spatial scalar on functional regression}
\usage{
\method{plot}{spfda}(object, q = 0.05)
}
\arguments{
\item{object}{mcmc result}

\item{q}{credible interval parameter. default value is 0.05}
}
\description{
\link{plot} is used to plot the estimated beta.
}
