% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.spfda.R
\name{print.summary.spfda}
\alias{print.summary.spfda}
\title{Summary the result of LSIRM model}
\usage{
\method{print}{summary.spfda}(x, ...)
}
\description{
\link{summary.spfda} is used to summary the result of Spatial scalar on functional regression.
}
