% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.spfda.R
\name{summary.spfda}
\alias{summary.spfda}
\title{Summary the result of LSIRM model}
\usage{
\method{summary}{spfda}(object, CI = 0.95)
}
\description{
\link{summary} is used to summary the result of Spatial scalar on functional regression.
}
